fix: removed internal types from transaction type
